feat(constants): introduce UserRole enum and update USER_ROLES structure
- Added a UserRole enum to define user roles (SUPER_ADMIN, ADMIN, EMPLOYEE, CUSTOMER) for better type safety.
- Refactored USER_ROLES to use the new enum, enhancing maintainability and clarity in role management.
